Support pkgname arrays
Previously only `pkgname`s with a single value were supported. The
PKGBUILD allows for `pkgname` with multiple values in an array.

The way this has been achieved is by reusing the existing `Pkgname`
struct with all its validation, but making it a `Vec<Pkgname>` inside
the `PkgData` struct.

If the input `PKGBUILD` has a `pkgname` assignment to an empty
array it will fail just as before, but the error message also
specifies now that it could be due to the `pkgname` variable being
an empty array.