Add support for pkgbase
The `pkgbase` key is an optional key whose only requirement is that
it should not start with a hyphen.

The documentation indicates that when this key is not set, it's
value is, by default, the first value of `pkgname`. However, the
library doesn't assign this default value, and it's up to the
applications to do so. This is because if the library were to
automatically assign to `pkgbase` the first `pkgname` value, the
application would have no way to tell if there was a `pkgbase`
defined in the `PKGBUILD` or not.

Doing it this way (leaving the default value setting up the
application) allows the application to know whether `pkgbase` was
defined in the `PKGBUILD` or not.